requirements.txt
flask
pyyaml
PyGithub
simplejson
